Changed to include tags
@lindsey-s @justineschott

This is the same logic as the CCK12 Wi-fi query change that I made a
pull request for in Ecto
(https://github.com/educationsuperhighway/ecto/pull/156/files). The
request is to be able to have dqt use a tag to say a district has
sufficient wifi or not, instead of using the connectivity questions.

Justine, can you QA to make sure this looks good? If you’d like to see
the Aha card: https://educationsuperhighway.aha.io/features/SAT-2368