Consortium Affilition
@justineschott Can you QA?

I took out the restriction that the applicant ID couldn’t be equal to
the district ID. It didn’t seem to make sense to have this restriction
in the case where we are saying the service provider is the Consortia.

I checked with @jamie-barnes and she thought this was a strange
restriction as well. I’m going to put in a PR in Ecto as well with the
same change and send to Meghan