Consultant Summary
From this card: https://www.pivotaltracker.com/story/show/141245261

One query summarizes the consultants by state, the other summarizes
them nationally. The goal was to break out the number of districts (and
types of districts) the consultants file for so that the district team
can do outreach through the state departments.

See QA Plan on the pivotal card